wrap int() around all indexes to force behavior

+++ rpihat/pimotorhat.py
@@ -145,17 +145,17 @@ class Raspi_StepperMotor:
         try:
             pwm_a = pwm_b = 0
             if (self.currentstep >= 0) and (self.currentstep < self.MICROSTEPS):
-                pwm_a = self.MICROSTEP_CURVE[self.MICROSTEPS - self.currentstep]
+                pwm_a = self.MICROSTEP_CURVE[int(self.MICROSTEPS - self.currentstep)]
                 pwm_b = self.MICROSTEP_CURVE[self.currentstep]
             elif (self.currentstep >= self.MICROSTEPS) and (self.currentstep < self.MICROSTEPS*2):
-                pwm_a = self.MICROSTEP_CURVE[self.currentstep - self.MICROSTEPS]
-                pwm_b = self.MICROSTEP_CURVE[self.MICROSTEPS*2 - self.currentstep]
+                pwm_a = self.MICROSTEP_CURVE[int(self.currentstep - self.MICROSTEPS)]
+                pwm_b = self.MICROSTEP_CURVE[int(self.MICROSTEPS*2 - self.currentstep)]
             elif (self.currentstep >= self.MICROSTEPS*2) and (self.currentstep < self.MICROSTEPS*3):
-                pwm_a = self.MICROSTEP_CURVE[self.MICROSTEPS*3 - self.currentstep]
-                pwm_b = self.MICROSTEP_CURVE[self.currentstep - self.MICROSTEPS*2]
+                pwm_a = self.MICROSTEP_CURVE[int(self.MICROSTEPS*3 - self.currentstep)]
+                pwm_b = self.MICROSTEP_CURVE[int(self.currentstep - self.MICROSTEPS*2)]
             elif (self.currentstep >= self.MICROSTEPS*3) and (self.currentstep < self.MICROSTEPS*4):
-                pwm_a = self.MICROSTEP_CURVE[self.currentstep - self.MICROSTEPS*3]
-                pwm_b = self.MICROSTEP_CURVE[self.MICROSTEPS*4 - self.currentstep]
+                pwm_a = self.MICROSTEP_CURVE[int(self.currentstep - self.MICROSTEPS*3)]
+                pwm_b = self.MICROSTEP_CURVE[int(self.MICROSTEPS*4 - self.currentstep)]
         except TypeError as te:
             print("Error indexing! self.currentstep ={0}, self.MICROSTEPS={1}".format(self.currentstep, self.MICROSTEPS))
             _,_, tb = sys.exc_info()
